Add auto conversion on numeric types*
**Context**

In a dynamic type language, which doesn't have proper type annotation
over `i8/i16/i32/i64`, `u8/u16/u32/u64` or `f32/f64`, it encodes a numeric
value simply for what it fits. Or in the case when we tranform a numeric
value from JSON to MsgPack, there is no way to specify the exact type
the value is.

**This commit**

Therefore, this commit updates the decoder to auto converts compatible
values and throws an error on value overflow and underflow.